requirements.txt
pytest
pylint
pandas
spacy
country_converter
rapidfuzz
boto3
python-dotenv
pandarallel